refactor(market-information): split observation_service by domain
Split the 888-line observation_service.go into four focused files:
- observation_record.go: single observation recording + validation helpers
- observation_batch.go: batch observation recording
- observation_query.go: retrieval and listing operations
- observation_service.go: shared helpers (error mapping, quality/proto converters)

Remove from knownOversizedFiles (167 lines, was 888).
